Rename librocc to libgeneric for now
