+++ librocc/generic.h
